backend: Move some common structs/functions into backend
Some structs that are used in ui need to be accessible from other
backends as well.

Speed, InterfaceSelection, (Cynthion/Device)Usability
Move/rename CynthionStop -> BackendStop
